requirements.txt
streamlit
pandas
openpyxl
pillow
torch
transformers
